third_party/blink/web_tests/wpt_internal/prerender/session-history-subframe-navigation.https.html

<!DOCTYPE html>
<title>Test history.length during a subframe navigation</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/session-history-test-util.js"></script>
<body>
  <script>
    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testSubframeNavigationInPrerender"),
        "Passed"
      );
    }, "Subframe navigation in prerender replaces the session entry");
  </script>
</body>
